refactor: took out logging in Google separately
